requirements.txt
datasets
pandas
groq
python-dotenv
pytest
flask
flask-cors
tqdm
streamlit>=1.35.0
altair>=5.0.0
